+++ lib/rack/request_police/storage/unit.rb
@@ -1,7 +1,7 @@
 module Rack
   module RequestPolice
     module Storage
-      class Unit < Struct.new(:method, :ip, :url, :time, :data)
+      class Unit < Struct.new(:method, :ip, :url, :time, :data, :headers)
       end
     end
   end